refactor(testsuite): Rename lints to lints_table
